move headers to separate directory
