+++ include/mb/aligned_ptr.h
