+++ src/pyramid.cpp
@@ -1,4 +1,4 @@
-#include "src/pyramid.h"
+#include "mb/pyramid.h"
 
 #ifdef PYR_DENOISE
 #define USE_SSE2
@@ -7,10 +7,10 @@
 
 #include <cmath>
 
-#include "src/aligned_ptr.h"
-#include "src/linux_overrides.h"
-#include "src/pnger.h"
-#include "src/threadpool.h"
+#include "mb/aligned_ptr.h"
+#include "mb/linux_overrides.h"
+#include "mb/pnger.h"
+#include "mb/threadpool.h"
 
 namespace multiblend {
