+++ src/tiff.cpp
@@ -1,4 +1,4 @@
-#include "src/tiff.h"
+#include "mb/tiff.h"
 
 #include <cstdint>
 #include <cstdlib>
